+++ sdk/core/iOS/MoEReactBridge/MoEngageInitializer.h
@@ -9,6 +9,7 @@
 #import <UIKit/UIKit.h>
 
 @protocol SFSafariViewControllerDelegate;
+@class MoEngageReactSDKInitializationConfig;
 #import <MoEngageSDK/MoEngageSDK.h>
 
 NS_ASSUME_NONNULL_BEGIN
@@ -31,6 +32,8 @@ NS_ASSUME_NONNULL_BEGIN
 /// @warning Make sure to call only one of the initialization methods available (either with plist OR with MOSDKConfig instance)
 /// @version 8.1.0 and above
 - (void)initializeDefaultSDKConfigWithState:(MoEngageSDKConfig*)sdkConfig withSDKState:(MoEngageSDKState)sdkState andLaunchOptions:(NSDictionary*)launchOptions;
+
+- (void)initializeInstance:(MoEngageReactSDKInitializationConfig*)reactConfig;
 @end
 
 NS_ASSUME_NONNULL_END