+++ sdk/core/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next Version
 - Support for JSONArray and JSONObject in Event & User Attributes.
+- Android
+  - `moe-android-sdk` version updated to `13.02.00`
+  - `inapp` version updated to `8.3.1`
 
 # 16-05-2024